spec: Update and simplify packaging for Fedora and add EL8 support
The packaging was a tad bit more complicated than it needed to be
for supporting Fedora and openSUSE equally.

This change simplifies it by bringing it up to date with latest
packaging practices in Fedora and adding the macro to enable the
Python dependency generator for EL8 so that dependencies are
correctly auto-generated when building a package for EL8 (with
dependencies from EPEL).

Signed-off-by: Neal Gompa <ngompa13@gmail.com>