Implement --repo-priority / --no-repo-priority
This commit provides a way to turn off the current behavior where sesdev adds
custom repos with an elevated priority.

Also, it drops the setting of elevated priority on repos hard-coded in
VERSION_OS_REPO_MAPPING. These repos are controlled by package maintainers and
are not expected to need elevated priority.

Fixes: https://github.com/rjfd/sesdev/issues/18
Signed-off-by: Nathan Cutler <ncutler@suse.com>