qa/health-ok.sh: wait for OSD nodes to show up
After a recent round of refactoring, it started taking a longer time for
the cluster to become aware of newly added OSD nodes. As a result,
number_of_nodes_actual_vs_expected_test was failing because the script
was going too fast. This commit causes the script to wait, if necessary,
for up to 5 minutes for the OSD node(s) to show up.

Signed-off-by: Nathan Cutler <ncutler@suse.com>