Implement feature "sesdev create makecheck"
Originally, this feature was envisioned as "sesdev make-check VERSION",
but during the course of implementation the following was found to be
more expedient:

    sesdev create makecheck \
        --os leap-15.2 \
        --ceph-repo https://github.com/ceph/ceph \
        --ceph-branch octopus

(The main reason is that "make check" is tied more to operating system
than to SES version.)

When the command is issued, sesdev will create and provision a single
VM, and inside that VM it will create an ordinary user and grant it
passwordless sudo privileges. That user will then clone the git repo,
check out the branch, and run install-deps.sh and run-make-check.sh
inside it.

Fixes: https://github.com/SUSE/sesdev/issues/82
Signed-off-by: Nathan Cutler <ncutler@suse.com>