seslib: force ses5 prometheus node to master
The DeepSea version that is shipped with SES5 does not have any
"prometheus" or "grafana" roles. It always deploys these services and
there is no easy way to prevent it from doing that.

For port-forwarding purposes, sesdev forces the grafana node to be
"master" in ses5 deployments, and it should have been doing this for
prometheus as well.

Fixes: https://github.com/SUSE/sesdev/issues/231
Signed-off-by: Nathan Cutler <ncutler@suse.com>