provision.sh: reinstall certain packages from the update repos
The SLE-12-SP3 Vagrant Box has newer versions of the python-base,
python-xml, libncurses5, and libncurses6 packages than are available in
the standard repos. This is a Bad Thing and, I might add, a bug in the
SLE-12-SP3 Vagrant Box. It also makes install-deps.sh fail. Work around
the bug by reverting these three packages to the latest version that is
in the repos.

Signed-off-by: Nathan Cutler <ncutler@suse.com>