qa: better MGR counting
I have not found any straightforward way of obtaining the number of
running MGRs from the "ceph status --format json" output. This commit
makes it more explicit how the MGR count is arrived at. First, a new
test is implemented, which fails if the mgrmap does not have "available:
false", since this seems like it would indicate a problem with the MGR
deployment. Second, iff that test passes, the total number of MGRs is
calculated as follows:

(1) if the mgrmap has "available: true" then "number of active MGRs"
    is 1, otherwise 0

(2) obtain the number of standby MGRs by consulting the mgrmap

(3) total number of MGRs = number of active MGRs + number of standby MGRs

Signed-off-by: Nathan Cutler <ncutler@suse.com>