qa/health-ok.sh: move health test to the end
In octopus and beyond, cephadm/orchestrator deploys the cluster
asynchronously, which means the cluster can start out in HEALTH_OK and
then go through HEALTH_ERR, HEALTH_WARN, etc. (as services are deployed)
and only reach the final HEALTH_OK later.

The right time to start waiting for HEALTH_OK is *after* we know all the
expected daemons are running -- not before.

Signed-off-by: Nathan Cutler <ncutler@suse.com>