Rename --deepsea-cli and --ceph-salt-deploy options to --salt
With this patch, when we want "salt ... state.apply" (SES7) or "salt-run
state.orch ..." (SES6, SES5), we simply give the option --salt. This is
much more straightforward than the previous method.

Note: this patch does *not* change the default behavior, which is to use
"ceph-salt deploy --non-interactive" (SES7) or "deepsea stage run ..."
(SES6, SES5). The patch only renames the options used to trigger the
non-default behavior.

Signed-off-by: Nathan Cutler <ncutler@suse.com>