seslib: more robust dashboard tunnelling
We assume that "master" will be an "admin" node, and hence that
"ceph mgr services" will work there.

This commit changes the logic so the output of "ceph mgr services" is
parsed in the Python code, instead of by piping the output to "jq" and
"sed" on the cluster node like we were doing before.

It also adds a regression test for the dashboard tunnelling feature,
which also asserts that at least *some* form of dashboard downstream
branding has been applied in SES5, SES6 and SES7.

Also take the opportunity to make "contrib/standalone.sh"
ShellCheck-clean.

Fixes: https://github.com/SUSE/sesdev/issues/269
Signed-off-by: Nathan Cutler <ncutler@suse.com>