tunnel dashboard: more accurate info message
When running

    sesdev tunnel DEP_ID dashboard

what happens is "ceph mgr services" is run via SSH on the node called
"master". The actual tunnel gets created to the node with the active
dashboard instance on it.

So, mentioning the master node here is confusing, because it seems to
imply that the active dashboard is on "master", which might not be true.

Signed-off-by: Nathan Cutler <ncutler@suse.com>