provision.sh: refactor PROVISIONED_NODES while loop
If all the "/tmp/ready-{{ node.name }}" files don't get created in the
first iteration of the loop, the scp fails and that causes the whole
script to fail because it is running under "set -e".

(This while loop was written before we had "set -e" at the top of the
script.)

Since we're touching the code of the loop anyway, take this opportunity
to refactor it.

Fixes: c0d5df13b20635ad8a7a0075f74b21800cd10ffc
Fixes: https://github.com/SUSE/sesdev/issues/278
Signed-off-by: Nathan Cutler <ncutler@suse.com>