provision.sh: enable autorefresh on all repos
We transiently hit the following error when running "zypper install",
even when network connectivity is OK:

Media source
'http://download.opensuse.org/distribution/leap/15.2/repo/oss/' does not
contain the desired medium

This is because the SLE/Leap repos are periodically updated and we race
with these updates. To reduce the probability of hitting the error, we
enable "autorefresh" on all zypper repos.

Fixes: https://github.com/SUSE/sesdev/issues/287
Signed-off-by: Nathan Cutler <ncutler@suse.com>