seslib: explicit error-exit when duplicate roles on a node
Before this commit, if a user gave more than one of any given role on
a node, the extra instances (beyond the first) were silently dropped.

This behavior might surprise a user who is expecting sesdev to be able
to tell the Cephadm Orchestrator to deploy, e.g., two mgr daemons on one
node (sesdev is not currently able to do this).

Now that we have "vet_configurations" and "KNOWN_ROLES", we can display
a nice error message in such a case.

Signed-off-by: Nathan Cutler <ncutler@suse.com>