makecheck: possibly prophylactically downgrade problem packages
Compiling Ceph requires a number of "libfoo-devel" RPMs. These are typically
shipped in a different SLE Module than their corresponding library RPMs.
Sometimes, an update to the library RPM ("libfoo1") reaches the IBS Vagrant Box
before the corresponding update to the "libfoo-devel" package reaches the
IBS repos. This state, which can last for days or even weeks, typically causes
install-deps.sh to fail on a zypper conflict.

Possibly prophylactically downgrade libraries known to have caused this
problem in the past.

Signed-off-by: Nathan Cutler <ncutler@suse.com>