contrib/standalone.sh: --no-stop-on-failure option
From time to time, various deployment types and configurations can
become unreliable, making it very difficult to get through an entire
"contrib/standalone.sh --full" run even when there are no bugs in the
sesdev branch under test.

Since "contrib/standalone.sh --full" typically runs over night, it can
be very inconvenient when issues unrelated to sesdev cause the script to
abort. This commit provides a "--no-stop-on-failure" option for times
when we want the script to run to completion, regardless of failures.

Signed-off-by: Nathan Cutler <ncutler@suse.com>