seslib: move boilerplate ssh options into a staticmethod
At the beginning, Deployment had just one ssh method so there was no
code duplication.

Then, the scp feature was added and the ssh boilerplate options got
duplicated.

Recently, the rsync feature was added, which uses the same ssh
boilerplate options. At that point, the same exact set of boilerplate
options appeared in three different locations.

If we ever wanted to modify the boilerplate options, we would have to
modify it in at least three places. To avoid that, move the boilerplate
options into a dedicated, static method.

Signed-off-by: Nathan Cutler <ncutler@suse.com>