Deployment: vet deployment ID before creation
It's not a good idea to allow users to specify a deployment ID that
would not be a valid hostname element [1], since the deployment ID
might be used as a hostname element.

[1] from hostname(7) - Linux manual page:

Each element of the hostname must be from 1 to 63 characters long and the
entire hostname, including the dots, can be at most 253 characters long.
Valid characters for hostnames are ASCII(7) letters from a to z, the digits
from 0 to 9, and the hyphen (-).

Signed-off-by: Nathan Cutler <ncutler@suse.com>