qa: implement rgw smoke test
For all nodes that are expected to have an RGW running, use "curl" to
send an "http(s)" request to the RGW port on the node. If an RGW really
is listening on the port, it will send back a response, the body of
which will contain valid XML.

Exploit this RGW behavior to implement a smoke test.

Fixes: https://github.com/SUSE/sesdev/issues/343
Signed-off-by: Nathan Cutler <ncutler@suse.com>