qa: add systemctl_list_units_test
Implement a systemctl-based test using ssh to run a script pre-installed
on all nodes.

The test script needs to know (or be able to look up) certain
characteristics of the cluster configuration. To address this, a file
(/home/vagrant/cluster.json) is created on each node. The cluster.json
file contains a data structure into which we place all the information
needed by tests that health-ok.sh triggers on an arbitrary node.

Signed-off-by: Nathan Cutler <ncutler@suse.com>