Add --devel/--product option and add-repo subcommand
The --devel/--product option (default: --devel) controls whether the
"devel" repo will be added to the nodes prior to deployment.

For SES{5,6,7}, the "devel" repo is Devel:Storage:{5,6,7}.0 Media1.

For {nautilus,octopus,pacific}, the "devel" repo is something like
"filesystems:ceph:...". For other create targets, the "devel" repo is
undefined and this PR is a noop.

This commit also implements an "add-repo" subcommand which can be used
to add either the "devel" repo or an arbitrary, "custom" repo to an
already-running cluster, and optionally update the cluster to newer (or
older) versions of packages contained in that repo at the same time.

Signed-off-by: Nathan Cutler <ncutler@suse.com>