seslib: no longer support nautilus deployment in Tumbleweed
Since Tumbleweed is now on Octopus, nautilus is not expected to be built
for Factory/Tumbleweed anymore in filesystems:ceph:nautilus.

Also, the "octopus" deployment target is expected to deploy something
very similar, if not identical, to SES7 on openSUSE Leap 15.2. So use
filesystems:ceph:octopus by default, instead of
filesystems:ceph:octopus:upstream.

Signed-off-by: Nathan Cutler <ncutler@suse.com>