qa/rgw: get RGW port from /etc/ceph/ceph.conf
In an SES5 single-node cluster, openattic and rgw are running on the
same node. In that case, DeepSea sets the RGW port to 8080 instead of
80. There might be other cases where the deployment tooling sets
a different port.

Make the RGW smoke test handle these cases better. It still only looks
in /etc/ceph/ceph.conf (not in the MON config store), but it's an
incremental improvement.

Signed-off-by: Nathan Cutler <ncutler@suse.com>