README.md: sesdev can be installed on SLE, too
sesdev can be installed and used on SUSE Linux Enterprise just as well
as on openSUSE. The README should make that clear.

Also, it's not necessary to explicitly write "sudo" everywhere. Users
are expected to be savvy enough to know what "run as root" means.

Finally, AFAICT bridge-utils is not needed for anything (it contains
a single executable - brctl - which has been obsoleted by iproute2).

Fixes: 8dcf52429dfd4a67b1dda19984b33d490f179f61
Signed-off-by: Nathan Cutler <ncutler@suse.com>