seslib: do not use CR container for octopus deployment
By default, we were using the CR RPMs/containers for the octopus
deployment, but that's no longer appropriate because we want users to
get the same code when they deploy "octopus" as when they deploy "ses7".

f67dbb87cfba4c9a10ff0f02720f325ede994a6e fixed this for the RPMs, but
neglected to make the same change for the container.

Fixes: f67dbb87cfba4c9a10ff0f02720f325ede994a6e
Signed-off-by: Nathan Cutler <ncutler@suse.com>