seslib/deployment: try to destroy the whole cluster at once
Vagrant knows how to destroy the whole cluster with a single command,
so try that first. If that fails, fall back to destroying each
node individually like we were doing before, except ignore errors.

This fixes the issue where a cluster that is expected to have nodes:
"master", "node1", "node2", and "node3" but in reality only has
two nodes "node2" and "node3" cannot be destroyed because sesdev
first tries to destroy "master" but this fails, causing sesdev to bail
out.

Signed-off-by: Nathan Cutler <ncutler@suse.com>