ses5: stop patching srv/salt/ceph/time/ntp/default.sls
This partially reverts f1fb650b2309e9f91098da02bff795238e6d95a2
which was monkey-patching srv/salt/ceph/time/ntp/default.sls to work
around a bug in the Salt that affects SES5. (That bug was preventing
sesdev from syncing the clocks.) In the meantime the following
commit:

https://github.com/SUSE/DeepSea/pull/1823/commits/d22232629d7d55387b7d6eaee7a510eced848d34

was merged into the SES5 DeepSea branch and started shipping to
customers, so the monkey-patch no longer applies cleanly (and is
no longer needed).

Signed-off-by: Nathan Cutler <ncutler@suse.com>