Change "version_os_repo_mapping" to "version_devel_repos"
I never liked "version_os_repo_mapping". In our SES team parlance,
these repos are collectively referred to as "devel repos", so the
name will be more descriptive, and less confusing, if we include the
term "devel repos" in it.

Of course, it's a bit of a long-shot to say that the team is used to
referring to, e.g., "filesystems:ceph:nautilus" as the "SES6 devel repo
for OBS", but that's effectively what it is. (The OBS equivalent
of the SES6 product repo is the Leap 15.1 base repo.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>