cephadm: allow preparation of cluster for manual deployment
Before this commit, the only way to prepare a cluster for manual Day
2 deployment would be to pass the `--stop-before-ceph-orch-apply`
option. But it seems weird to be forced to assign fictional roles just
to get a cluster with X nodes that has just been bootstrapped.

With this commit, machines with no roles at all [1] will still get added
to the cluster. Later, the user can SSH into the cluster and use these
machines in manual Day 2 deployment operations.

[1] For example: --roles="[master],[mon,mgr,bootstrap],[],[],[]"

Signed-off-by: Nathan Cutler <ncutler@suse.com>