deepsea: properly recognize deprecated ganesha role
As far as sesdev is concerned, the role "ganesha" should not be used,
because it has been renamed to "nfs". However, this can be confusing.
First, we had to keep the "ganesha" role around to avoid breaking
running clusters. Second, DeepSea also has a concept of "roles" and, as
far as DeepSea is concerned, the role is still called "ganesha".

This commit tries to bring some order to the chaos by:

1. tolerating "ganesha" (as a synonym for "nfs") in clusters
   that use DeepSea as a deployment tool
2. raising an exception, with the effect of displaying a helpful message,
   if someone unknowingly specifies a "ganesha" role for a post-nautilus
   cluster.

Fixes: 6d18054d774faf8d4aea3802b4bfe666c9189e91
Signed-off-by: Nathan Cutler <ncutler@suse.com>