ceph-salt: enable user to control which nodes get "admin" role
The documentation states that the only node that **must** have an
"admin" role is the bootstrap node. On all other nodes, including the
master node, "admin" is optional.

I realized that our previous approach, which was to apply the "admin"
role to all nodes without exception, might not be ideal for realistic
testing. According to the SES7 manual, this configuration is not
recommended.

So, this commit makes it possible for the user to specify which nodes
get the "admin" role when deploying {octopus,ses7,pacific} clusters,
with the caveat that then "admin" role must be given on nodes that
have the "bootstrap" and/or the "master" role. (On the "master" node,
"admin" is not strictly required, but it seems safe to assume that most
real-life deployments will have it, since it's natural to want to run
ceph CLI commands on the master node.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>