deployment: never use "admin" as the name of a node
A long time ago, we used to have a single "admin" node and assigned
the string "admin" as its short hostname. Later, we changed this to
"master" and, in doing so, I added a piece of backward compatibility
code so the change would not break running clusters.

Now this backward compatibility code is making it difficult to bring
back the "admin" role. Since it's extremely unlikely that anyone still
has a cluster running that would need this backward compatibility code,
this commit is dropping it.

Signed-off-by: Nathan Cutler <ncutler@suse.com>