ceph_salt_deployment: explicitly create mds service
Until now we were relying on "ceph fs volume create" to create the mds
service (we use the word "service" here in the cephadm sense), but
apparently it no longer does that.

Explicitly deploy the mds service/daemons by compiling a service spec
file and running "ceph orch apply" on it.

Signed-off-by: Nathan Cutler <ncutler@suse.com>