contrib/standalone.sh: adapt to create always returning 0
In order to work around a regression in vagrant, we were forced in
56ca70ad243e14d849cd289858cff02746cac59a to make "sesdev create" always
return exit status 0. This had the side effect of breaking failure
detection in contrib/standalone.sh.

Fortunately, "sesdev qa-test ..." still returns non-zero on failure, so
we can take advantage of that to detect failed deployments.

(Later we will need to capture the output from "sesdev create" and grep
the last few lines for strings indicating a failure.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>