ceph_salt_deployment: always wait for OSDs to appear
During integration testing we found that ceph_salt_deployment.sh always
waits for OSDs to appear, except in one test case:

    sesdev create ses7 --roles [admin,master,bootstrap,storage,mon,mgr]

In this case, since there are no gateways, the "wait for OSDs to appear"
routine was not triggered.

This commit revamps the conditionals so the provisioning script always
waits for the OSDs to appear (if there are any OSDs).

Signed-off-by: Nathan Cutler <ncutler@suse.com>