ceph-salt: deploy and smoke-test node-exporter
With this PR, nodes that have the "node-exporter" role will get a
"node-exporter" service deployed. The "node-exporter" role is only
available in {octopus, ses7, pacific} (in {nautilus, ses6}, DeepSea
deploys the node-exporter together with Prometheus).

The commit also implements a node-exporter smoke test.

With this commit, sesdev can now deploy the full monitoring stack in
{octopus,ses7,pacific}.

Fixes: https://github.com/SUSE/sesdev/issues/354
Signed-off-by: Nathan Cutler <ncutler@suse.com>