qa: wait for grafanas to show up before counting them
As we know, cephadm deployments are asynchronous. In practice, this
means you issue a "ceph orch apply" command and it returns quickly,
but you have to wait for the results.

Until now, the preceding IGW wait took long enough that any remaining
daemons got deployed in the meantime. But recently we saw the following
failure in the Jenkins CI environment:

    master: WWWW: number_of_services_expected_vs_orch_ls_test
    master: MGR services (orch ls/expected): 1/1
    master: MON services (orch ls/expected): 1/1
    master: MDS services (orch ls/expected): 1/1
    master: RGW services (orch ls/expected): 1/1
    master: NFS services (orch ls/expected): 1/1
    master: IGW services (orch ls/expected): 1/1
    master: Prometheus services (orch ls/expected): 1/1
    master: Grafana services (orch ls/expected): 0/1
    master: alertmanager services (orch ls/expected): 1/1
    master: node-manager services (orch ls/expected): 1/1
    master: ERROR: Detected disparity between expected number of services and "ceph orch ls"!
    master: WWWW: number_of_services_expected_vs_orch_ls_test: FAIL

which we address by adding a grafana wait to the existing list of waits.

Signed-off-by: Nathan Cutler <ncutler@suse.com>