ceph-salt: no cephadm for nodes with no sesdev roles
In order to be able to simulate administration tasks like adding nodes to
a cluster, we need to make it easy for sesdev users to create deployments that
have "extra" nodes.

Before this change, we had to specify the "extra" node's roles like this:

    [client]

Now nodes with no roles at all will be treated the same way when it comes to
assigning the "cephadm" ceph-salt role.

Signed-off-by: Nathan Cutler <ncutler@suse.com>