zypper: be more careful when installing systemd-coredump
Due to the way Kiwi builds SLE Vagrant Boxes, the versions of packages
pre-installed in the Vagrant Box may be newer than the versions of the same
packages in the "channels" (i.e., in the zypper repos that are added after the
VMs start).

Work around this by force-installing systemd, udev, etc. from the repos.

This has the drawback of putting the system into a state where "zypper ps -s"
reports that it needs a reboot.

Fixes: https://github.com/SUSE/sesdev/issues/575
Signed-off-by: Nathan Cutler <ncutler@suse.com>