qa: add assert_reboot_not_needed test
As part of qa-test, run "zypper ps -s" on all nodes and evaluate the output.

If any node is in need of reboot, abort the test and instruct the user that
it would be better if the cluster were rebooted before running qa-test.

Since rebooting the cluster before running qa-test would have to be done
manually, failure of this test currently only results in a "soft failure"
(warning), not a hard failure.

Signed-off-by: Nathan Cutler <ncutler@suse.com>