apply custom configs after provisioning
I've been using this for a while now and think it might also be useful
for others. Although, it would also be useful for me to have it merged.
It's not a full featured solution, it just enables to comfortably and
automatically copy files to root and run an arbitrary script on each
host on deployment. The custom provisioning can also be triggered
manually after deployment for all VMs of a deployment or just one.

Signed-off-by: Patrick Seidensal <pseidensal@suse.com>