deployment: run "vagrant up" with --no-destroy-on-error
Up to Vagrant 2.2.9, --no-destroy-on-error was the default and everything was
fine.

Then Vangrant 2.2.10 changed the default to --destroy-on-error. Since I was
ignorant of this option, I worked around it by merging
55314a49fb94c08a7681b72398eee33aaf9edef0 which ensured that the provisioner
script exits with status code 0 even when there is an error (that tricked
Vagrant into thinking the script succeeded, so it did not destroy the VMs when
the script failed).

Now that I know about this option, it makes sense to run with
--no-destroy-on-error since we don't want Vagrant to destroy VMs during
provisioning in any case (since it is assumed that we want to go in and debug to
determine why the provisioner failed).

Signed-off-by: Nathan Cutler <ncutler@suse.com>