Fix --stop-before-cephadm-bootstrap switch
`--stop-before-cephadm-boostrap` didn't work because the argument wasn't
part of the code that reads the arguments and creates the context out of
that. The context is then used to generate the templates (Vagrantfile
and deployment scripts). This means that the switch never reached the
conditions in the Jinja templates and was therefore technically ignored.

Signed-off-by: Patrick Seidensal <pseidensal@suse.com>