templates/zypper: ensure rsync is installed on all VMs
rsync must be installed on all VMs for the user-provision feature to work.
Often rsync gets pulled in as a dependency of some other package, but there
might be circumstances (?) when this doesn't happen.

This commit explicitly forces installation of rsync on all VMS.

In general, installing packages everywhere unconditionally might hide some
undeclared dependency bugs, but this seems unlikely in the case of rsync.

Fixes: https://github.com/SUSE/sesdev/issues/608

Signed-off-by: Nathan Cutler <ncutler@suse.com>