pycodestyle: Fix issue with yaml.load (use safe_load)
Replace `yaml.load` with `safe_load` to fix pycodestyle issue and ensure
using tests that `safe_load` is capable of loading the `config.yaml`
file correctly, including the ability to use the YAML reference (which
is not explicitly documented in the sesdev docs but part of YAML).

`yaml.load` has been to be used without a Loader. However it is unsafe
and has been used way too frequently, although the documentation of
PyYAML has always stated that it is unsafe.

Fixes: https://github.com/SUSE/sesdev/issues/618

Signed-off-by: Patrick Seidensal <pseidensal@suse.com>