provision.sh: zealously remove bogus loopback assignments from /etc/hosts
The previous implementation only removed bogus loopback assignments from
the first line of /etc/hosts, and expected a single tab character between
each field when doing the replacement.  I've since found with vagrant 2.2.19
deploying SES6 on SLE 15 SP1, that the loopback assignment is in the middle
of the file, and uses spaces, not tabs.

This change makes provision.sh remove bogus loopback assignments wherever
they are in the file, and it doesn't care about what kind of whitespace is
used.

Signed-off-by: Tim Serong <tserong@suse.com>