deploy: exclude dedicated master from disks
If a node has only the roles `master` and `admin`, it should not be
provisioned with disks when the command line option `--num-disks` is
given.
There was a check in place, but it was broken because the node with the
`master` role must also have the `admin` role and the check did not
account for that.

fixes: #450