feature: generate maintentance test reports
- Add command for generating maintenance test report data on a running
  cluster.

When testing maintenance updates for various packages, certain
informations need to be collected and entered into the final report.
Thus far, the information was largely collected by hand from various
commands run against the cluster.

This change adds a new sub command, which aggregates this information
and prints it ready for copy-and-paste to the console.
Included in this information is
 - The command and configuration used to create the test cluster
 - Basic information about the resulting cluster layout
 - A list of packages and installation statuses from each relevenat
   repo on each host
 - The Ceph cluster status

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>