feature: create test deployment from maintenance id
- Create a test deployment with auto-generated settings from a
  maintenance id

Maintenance update package tests rely on a consisted test environment to
produce reliable results. This change codifies the settings used to
generate such an environment, so that operator errors are minimized.
This convenience command takes a maintenance id of the form
'S:M:xxxxx:yyyyyy' and generates a cluster with the right additional
repos and settings for reliable testing.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>