install: install updates from custom repos
Install updates from custom repos, if any are available, for all
packages that are installed before starting the installation procedure
for ceph.
This ensures that when custom repos provide e.g. new versions of salt
or zypper those versions are installed before the cluster is installed.
This functionality is useful for maintenance tests against packages used
during the installation of ceph.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>