maintenance report: add list of repos
Add list of configured repos to the maintenance report. This is useful
to trace back which repo url a specific package listed in the same
maintenance report came from. It will also be useful in the future to
ensure a package update test actually installed the expected packages
and not packages from a different repo.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>