makecheck: do not su in .bashrc for remote node unless ssh tty
For makecheck deployment there's been printing a welcome message
and switching to sesdev user from root login when one attempts
`sesdev ssh ...` for jumping into the node.

However, we should not do 'su - sesdev' and output any garbage
messages if only we are not in ssh tty, because it makes impossible
to run commands over ssh correctly once node is deployed.
This is crucial for such sesdev commands as 'stop', 'report', etc.
which are using ssh under the hood.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>