qa: Don't install ceph-common if already installed on master
This is related to https://github.com/SUSE/sesdev/issues/672.
When running `sesdev create ses7p --qa-test`, the
make_salt_master_an_admin_node_test function will try to install
ceph-common.  If we're suffering from subtle version drift
between ceph-common and ceph-base from different repos, this
will fail with something like:

  Problem: the installed ceph-base-16.2.9.536+g41a9f9a5573-150300.3.3.1.x86_64 requires 'ceph-common = 16.2.9.536+g41a9f9a5573-150300.3.3.1', but this requirement cannot be provided
   Solution 1: deinstallation of ceph-base-16.2.9.536+g41a9f9a5573-150300.3.3.1.x86_64
   Solution 2: do not install ceph-common-16.2.9.536+g41a9f9a5573-150300.6.3.1.x86_64
   Solution 3: break ceph-base-16.2.9.536+g41a9f9a5573-150300.3.3.1.x86_64 by ignoring some of its dependencies

This is because zypper is trying to upgrade us to a newer version
of ceph-common from a different repo.  We can avoid this problem
by not trying to install ceph-common if it's already present.

Signed-off-by: Tim Serong <tserong@suse.com>