k3s/longhorn: use LVM to support additional storage disks
This is a really simple use of LVM to just create a linear logical volume
over `/dev/vd[b-z]`, then mount that for use by Longhorn.  The default is
still to deploy with only one additional disk, but at least now if more
than one disk is requested, we'll actually do something with the extra
disks.

Signed-off-by: Tim Serong <tserong@suse.com>