Jenkinsfile.integration: collect more artifacts
In one recent CI run, the salt minion never returned and
the salt minion log and cephadm.out log just stopped dead
immediately after `cephadm bootstrap` was invoked.  This
commit collects `ps` and `journalctl` output in the hope
that this will help with further debugging if we hit this
issue again.

Signed-off-by: Tim Serong <tserong@suse.com>