README.md: drop mention of packages
The packages on OBS are very very old, and we haven't cut a new
release for a long time.  In order to avoid confusion (given sesdev
is not currently widely used), I've disabled package builds on OBS
and removed the mention of the packages from the README in favour
of just running from source.  If the situation changes in future
we can always re-enable builds on OBS and revert this change.

Fixes: #705
Signed-off-by: Tim Serong <tserong@suse.com>